requirements.txt
asgiref==3.7.2
Django==4.2.2
django-environ==0.11.2
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
packaging==24.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
